--- src/components/Fallbacks/CartAction/index.tsx
@@ -1,7 +1,0 @@
-import style from './style.module.css';
-
-function CartAction() {
-  return <div className={style.cartAction}>오류가 발생했습니다. 잠시 후 다시 시도해 주세요.</div>;
-}
-
-export default CartAction;
