+++ src/components/Fallbacks/CartActionError/style.module.css
@@ -1,18 +1,21 @@
-.cartAction {
+.cartActionError {
+  display: flex;
   position: fixed;
   top: var(--header-height);
   left: calc((100vw - var(--desktop-width)) / 2);
+  justify-content: center;
+  align-items: center;
   z-index: var(--header-z-index);
   background-color: var(--pink);
   padding: 0.75rem 4.8125rem;
   width: var(--desktop-width);
-  height: 2.5rem;
-  text-align: center;
 }
 
 @media screen and (max-width: 768px) {
-  .cartAction {
+  .cartActionError {
     left: 0;
+    flex-direction: column;
+    gap: 0.5rem;
     width: var(--mobile-width);
   }
 }
