+++ src/components/index.tsx
@@ -3,3 +3,4 @@ export { default as Header } from './Header';
 export { default as Layout } from './Layout';
 export { default as Dropdown } from './Dropdown';
 export { default as IntersectionObserverArea } from './IntersectionObserverArea';
+export * from './Fallbacks';
